refactor: Morpheusのvariant→src/altマップをmorpheusAssetsへ抽出
MorpheusImageとMorpheusAvatarで共有する単一出典に。MorpheusImageは
MorpheusImageVariantを再exportし既存importを非破壊。
